+++ art/__init__.py
@@ -5,7 +5,7 @@ from .art import aprint, art, randart
 from .art import tprint, tsave, text2art
 from .art import decor
 from .art import get_font_dic, set_default, help_func, art_list, font_list, decor_list
-from .art_param import ART_VERSION, FONT_NAMES, ART_NAMES, DECORATION_NAMES, DEFAULT_FONT
-from .art_param import ART_COUNTER, FONT_COUNTER, DECORATION_COUNTER
-from .art_param import NON_ASCII_ARTS, NON_ASCII_FONTS, ASCII_FONTS, ASCII_ARTS
+from .params import ART_VERSION, FONT_NAMES, ART_NAMES, DECORATION_NAMES, DEFAULT_FONT
+from .params import ART_COUNTER, FONT_COUNTER, DECORATION_COUNTER
+from .params import NON_ASCII_ARTS, NON_ASCII_FONTS, ASCII_FONTS, ASCII_ARTS
 __version__ = ART_VERSION
