+++ art/tests/test.py
@@ -4734,7 +4734,7 @@ True
 >>> text2 == text2_copy
 True
 >>> from art.art import font_size_splitter
->>> from art.art_param import FONT_MAP,RANDOM_FILTERED_FONTS
+>>> from art.params import FONT_MAP,RANDOM_FILTERED_FONTS
 >>> font_dicts = font_size_splitter(FONT_MAP)
 >>> len(font_dicts["small_list"])>0
 True
