+++ otherfile/art_decor_check.py
@@ -2,9 +2,9 @@
 """Art and decors duplication and UTF-8 compatibility testing script."""
 import sys
 import art
-from art.art_param import DECORATIONS_MAP as Decor_Dict
+from art.params import DECORATIONS_MAP as Decor_Dict
 from art.art_dic import art_dic as Art_Dict
-from art.art_param import NON_ASCII_ARTS
+from art.params import NON_ASCII_ARTS
 from font_wizard import is_utf8, is_ascii
 
 Failed1 = 0
